refactor: extract settings feature
